webapp/internal/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
Flask-Discord==0.1.64
bcrypt==4.0.1
Werkzeug==2.3.7